refactor(rapidfort): move marker helpers to leaf package and inline matcher-selection accessor
