+++ grype/pkg/zarf_provider.go
@@ -12,6 +12,7 @@ import (
 	"github.com/klauspost/compress/zstd"
 
 	"github.com/anchore/grype/grype/distro"
+	"github.com/anchore/grype/grype/rapidfort"
 	"github.com/anchore/grype/internal/log"
 	"github.com/anchore/syft/syft/format/syftjson"
 	"github.com/anchore/syft/syft/sbom"
@@ -57,7 +58,7 @@ func zarfProvider(userInput string, config ProviderConfig, applyChannel func(*di
 				Path: archivePath,
 			},
 		},
-		IsRapidFortImage: hasRapidFortMarkerInSBOM(s),
+		IsRapidFortImage: rapidfort.HasMarkerInSBOM(s),
 	}
 
 	return packages, ctx, s, nil
