+++ grype/rapidfort/marker_test.go
@@ -1,4 +1,4 @@
-package pkg
+package rapidfort
 
 import (
 	"testing"
@@ -9,7 +9,7 @@ import (
 	"github.com/anchore/syft/syft/sbom"
 )
 
-func TestHasRapidFortMarkerInResolver(t *testing.T) {
+func TestHasMarkerInResolver(t *testing.T) {
 	tests := []struct {
 		name     string
 		resolver file.PathResolver
@@ -27,19 +27,19 @@ func TestHasRapidFortMarkerInResolver(t *testing.T) {
 		},
 		{
 			name:     "resolver containing the marker path returns true",
-			resolver: file.NewMockResolverForPaths("/etc/os-release", RapidFortMarkerPath),
+			resolver: file.NewMockResolverForPaths("/etc/os-release", MarkerPath),
 			expected: true,
 		},
 	}
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			assert.Equal(t, test.expected, hasRapidFortMarkerInResolver(test.resolver))
+			assert.Equal(t, test.expected, HasMarkerInResolver(test.resolver))
 		})
 	}
 }
 
-func TestHasRapidFortMarkerInSBOM(t *testing.T) {
+func TestHasMarkerInSBOM(t *testing.T) {
 	tests := []struct {
 		name     string
 		sbom     *sbom.SBOM
@@ -66,8 +66,8 @@ func TestHasRapidFortMarkerInSBOM(t *testing.T) {
 		{
 			name: "catalog containing the marker path returns true",
 			sbom: &sbom.SBOM{Artifacts: sbom.Artifacts{FileMetadata: map[file.Coordinates]file.Metadata{
-				{RealPath: "/etc/os-release"}:   {},
-				{RealPath: RapidFortMarkerPath}: {},
+				{RealPath: "/etc/os-release"}: {},
+				{RealPath: MarkerPath}:        {},
 			}}},
 			expected: true,
 		},
@@ -75,7 +75,7 @@ func TestHasRapidFortMarkerInSBOM(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			assert.Equal(t, test.expected, hasRapidFortMarkerInSBOM(test.sbom))
+			assert.Equal(t, test.expected, HasMarkerInSBOM(test.sbom))
 		})
 	}
 }
